Added a specific view for the postcode screen. This should seperate out the main content view that will store navigation views and the views they will display too
